requirements.txt
farm-haystack
unstructured
python-pptx
langchain 
tiktoken
langchainhub
pymilvus
llama-index
transformers
img2table
openai
asyncpg
psycopg2-binary
sqlalchemy[asyncio]
pyarrow 
pymilvus==2.3.4
python-dotenv
openpyxl
numpy==1.23.5